# apps/api/modules/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Nuage API
  description: |
    Minimalist, developer-friendly cloud storage for the Facile Suite.

    ## Authentication

    Most endpoints require authentication via one of these methods:

    - **Bearer token**: `Authorization: Bearer {token}` header
    - **Query parameter**: `?token={token}` fallback for contexts where headers are impractical
    - **HTTP Basic Auth**: WebDAV only — username is your email, password is your API token
    - **OIDC**: OpenID Connect flow when enabled by the instance administrator

    ## Rate Limiting

    All endpoints are rate-limited to **100 requests per minute per IP**.
    Exceeding the limit returns `429 Too Many Requests`.

    ## File Size Limits

    - Standard uploads: 100 MB max
    - Avatar uploads: 6 MB max
    - Chunked uploads: no per-request limit (total size declared at init)
  version: 1.0.0
  contact:
    name: Facile Studio
    url: https://facile.studio
  license:
    name: MIT
servers:
  - url: https://nuage.facile.studio
    description: Production
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /auth/register:
    post:
      operationId: register
      summary: Register a new account
      description: |
        Create a new user account with email and password.
        Disabled when `sso_only` is true.
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
            example:
              email: "alice@example.com"
              password: "hunter2hunter2"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                    format: uuid
                  token:
                    type: string
              example:
                user_id: "550e8400-e29b-41d4-a716-446655440000"
                token: "nua_tok_a1b2c3d4e5f6..."
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "conflict"
                message: "email already registered"
  /auth/login:
    post:
      operationId: login
      summary: Log in
      description: |
        Authenticate with email and password. Returns a bearer token.
        Disabled when `sso_only` is true.
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
            example:
              email: "alice@example.com"
              password: "hunter2hunter2"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                    format: uuid
                  token:
                    type: string
              example:
                user_id: "550e8400-e29b-41d4-a716-446655440000"
                token: "nua_tok_a1b2c3d4e5f6..."
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        API token passed as `Authorization: Bearer {token}`.
        Alternatively, pass `?token={token}` as a query parameter.
    basicAuth:
      type: http
      scheme: basic
      description: |
        WebDAV only. Username is your email, password is your API token.
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "invalid_argument"
            message: "validation failed"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "unauthenticated"
            message: "missing or invalid token"
...
security:
  - bearerAuth: []